홈페이지와 리스트페이지,카드와 개별피드 링크연결
