+++ main/src/pages/list/ListPage.jsx
@@ -3,6 +3,7 @@ import CardListManager from '../../components/CardListManager'
 import styled from 'styled-components'
 
 const Container = styled.div`
+  background-size: contain;
   background: var(--Grayscale-20, #f9f9f9);
 `
